[refactor] #54 - 코드 컨벤션 관련 PR 수정 요청사항 반영

+++ Carmunication/Models/KeychainItem.swift
@@ -154,9 +154,12 @@ struct KeychainItem {
      You should store the user identifier in your account management system.
      */
     static var currentUserIdentifier: String {
+        guard let bundleIdentifier = Bundle.main.bundleIdentifier else {
+            return ""
+        }
         do {
             let storedIdentifier = try KeychainItem(
-                service: Bundle.main.bundleIdentifier!,
+                service: bundleIdentifier,
                 account: "userIdentifier"
             ).readItem()
             return storedIdentifier
@@ -166,9 +169,12 @@ struct KeychainItem {
     }
 
     static func deleteUserIdentifierFromKeychain() {
+        guard let bundleIdentifier = Bundle.main.bundleIdentifier else {
+            return
+        }
         do {
             try KeychainItem(
-                service: Bundle.main.bundleIdentifier!,
+                service: bundleIdentifier,
                 account: "userIdentifier"
             ).deleteItem()
         } catch {
